+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * minimum Python version is 3.10
 * minimum lxml version is 5
+* improved support for XML namespaces
 * (internal) packaging updates
 
 ## 0.2.0 – 2024-09-29